Use floats in gain and decibel conversion routines

+++ include/math_utils.h
@@ -127,14 +127,14 @@ constexpr T1 left_shift_signed(T1 value, T2 amount)
 	return static_cast<T1>(shifted);
 }
 
-inline double decibel_to_gain(const double decibel)
+inline float decibel_to_gain(const float decibel)
 {
-	return pow(10.0, decibel / 20.0);
+	return powf(10.0f, decibel / 20.0f);
 }
 
-inline double gain_to_decibel(const double gain)
+inline float gain_to_decibel(const float gain)
 {
-	return 20.0 * log(gain) / log(10.0);
+	return 20.0f * logf(gain) / logf(10.0f);
 }
 
 // A wrapper to convert a scalar gain to a percentage.
